requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.1
async-timeout==5.0.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
cytoolz==1.1.0
Django==5.2.11
django-webpush==0.3.6
eth-hash==0.7.1
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.8.0
grpcio==1.78.0
hiero-sdk-python==0.2.1
http-ece==1.2.1
idna==3.11
intasend-python==1.1.2
multidict==6.7.1
parsimonious==0.10.0
pillow==12.1.1
propcache==0.4.1
protobuf==6.33.5
py-vapid==1.9.4
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
pywebpush==2.3.0
regex==2026.2.28
requests==2.32.5
sqlparse==0.5.5
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
